requirements.txt
discord.py
pymongo
dnspython
ro-py
asyncio
discord-pretty-help
discord.py-embed
requests
google